sql: refactor join predicates
refactor out join predicate definitions, implementations into separate
file.
